Upgrading to the latest YUI compressor
The latest YUI compressor includes a fix for a bug with media-queries in CSS which was causing problems on my site. I've also left the YUI filename so that it includes the version of YUI that is being used so it is easier to instantly see what version of the software you are using.